Solved unknown radius problem, Refactor the code so that codes for update is in step now

+++ app/src/main/java/com/example/chenrui/game1942application/Pos.java
@@ -14,8 +14,8 @@ public class Pos {
         this.y = y;
     }
 
-    public double getDistance(Pos pos){
-        return Math.sqrt(Math.pow(pos.x - x, 2) + Math.pow(pos.y - y, 2));
+    public float getDistance(Pos pos){
+        return (float) Math.sqrt(Math.pow(pos.x - x, 2) + Math.pow(pos.y - y, 2));
     }
 
     /*
